Updated car speeds

+++ include/Car.h
@@ -6,15 +6,15 @@
 class Car : public Entity {
 private:
 protected:
-	int direction; // 1 for right, -1 for left
+	float direction; // 1 for right, -1 for left
 public:
 	Car(void);
 	~Car(void);
 
 	virtual void Move(float deltaTime);
 
-	void Direction(int set_direction);
-	int Direction();
+	void Direction(float set_direction);
+	float Direction();
 
 };
